:computer: moves player and video js modules to lib
